[arch]: feat: start impling regfile
